Move champ efficiency calculation to separate class
